file_manager: Fix signed vs unsigned comparison warning
shell-encryption and private_membership leaked some configs
which disable some default warnings.

In order to re-enable all these warnings for chrome/browser/*
in this CL crrev.com/c/3303510, all warnings need to be fixed.

This CL is fixing the comparison between two integers for them
to be from the same exact type (i.e. integer).

Bug: chromium:1212837, b:201763017
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3327145
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951220}

+++ chrome/browser/ash/file_manager/copy_or_move_io_task.cc
@@ -270,7 +270,7 @@ void CopyOrMoveIOTask::GotFreeDiskSpace(int64_t free_space) {
 
   // Move operations that are same-filesystem do not require disk space.
   if (progress_.type == OperationType::kMove) {
-    for (int i = 0; i < source_sizes_.size(); i++) {
+    for (size_t i = 0; i < source_sizes_.size(); i++) {
       if (!IsCrossFileSystem(progress_.sources[i].url,
                              progress_.destination_folder)) {
         required_bytes -= source_sizes_[i];
